# static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Whiteflag API",
        "version": "1.0.0-alpha",
...
        "x-logo": {
            "url": "wflogo.png",
            "altText": "Whiteflag"
        }
    },
...
    "servers": [
        {
            "url": "http://{hostname}:{port}/",
            "variables": {
                "hostname": {
                    "default": "localhost",
                    "description": "Configurable hostname"
                },
                "port": {
                    "default": "5746",
                    "description": "Configurable port number"
                }
            }
        }
    ],
    "security": [
        {
            "description": "Basic HTTP authorization",
            "type": "http",
            "scheme": "basic"
        }
    ],
    "paths": {
        "/messages": {
            "get": {
                "tags": [ "Messages" ],
                "description": "Returns an array of messages in the primary datastore.",
                "operationId": "getMessages",
                "parameters": [
                    { "$ref": "#/components/parameters/transactionHash" },
                    { "$ref": "#/components/parameters/blockchain" }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/wfMessages"
                    },
                    "400": {
                        "$ref": "#/components/responses/error400"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
...
            "transactionHash": {
                "description": "The hash of a blockchain transaction",
                "name": "transactionHash",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            "blockchain": {
                "description": "The name of a blockchain",
                "name": "blockchain",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                }
            }
        },
        "schemas": {
            "wfMessage": {
                "title": "Whiteflag Message",
                "description": "Whiteflag message with MetaHeader as used by the API",
                "version": "1",
                "type": "object",
                "required": [ "MetaHeader", "MessageHeader", "MessageBody" ],
                "additionalProperties": false,
                "properties": {
                    "MetaHeader": {
                        "$ref": "protocol/metaheader.schema.json"
                    },
                    "MessageHeader": {
                        "$ref": "protocol/message.schema.json#/properties/MessageHeader"
                    },
                    "MessageBody": {
                        "$ref": "protocol/message.schema.json#/properties/MessageBody"
                    }
                }
            },
            "responseBodyMetaObject": {
                "title": "API response metadata",
                "description": "Meta data about the processing of the request",
                "type": "object",
                "properties": {
                    "additionalProperties": true,
                    "operationId": {
                        "description": "Unique string defined by the OpenAPI defintion used to identify the operation",
                        "type": "string"
                    },
                    "request": {
                        "description": "Details of the request on the endpoint",
                        "type": "object",
                        "properties": {
                            "client": {
                                "description": "The ip address of the client that made the request",
                                "type": "string"
                            },
                            "method": {
                                "description": "The HTTP method used to make the request",
                                "type": "string"
                            },
                            "endpoint": {
                                "description": "The endpoint on which the request has been made",
                                "type": "string"
                            }
                        }
                    },
                    "info": {
                        "description": "Information about how the data was processed",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "warnings": {
                        "description": "Warnings generated when processing the request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "errors": {
                        "description": "Errors generated when processing the request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "responseBodyErrorObject": {
                "title": "API response errors",
                "description": "Errors describing why the request could not succesfully be fulfilled",
                "type": "array",
                "items": {
                    "type": "string"
                }
            },
            "responseBodyErrors": {
                "type": "object",
                "properties": {
                    "meta": {
                        "$ref": "#/components/schemas/responseBodyMetaObject"
                    },
                    "errors": {
                        "$ref": "#/components/schemas/responseBodyErrorObject"
                    }
                }
            }
        },
...
        "responses": {
            "error400": {
                "description": "Invalid request, typically because of a malformed syntax or protocol error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
...
            "error500": {
                "description": "Internal error preventing the API to process the request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
...
            "wfMessages": {
                "description": "Successfully processed Whiteflag message query and returning an array of messages",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "description": "Array with resulting messages from a database query",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/wfMessage"
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        }
    }
}